refactor(training): schemas split + api/errors bases (Fase 5.6–5.8)
Fase 5.6 — split schemas.py (569L, 52 classes) → schemas/ pacote:
  - sessions.py  (TrainingSessionOut, CreateTrainingSessionIn, UpdateTrainingSessionIn, TransitionOut)
  - blocks.py    (SessionBlockOut, AddSessionBlockIn, UpdateSessionBlockIn, ReorderSessionBlocksIn)
  - attendance.py (AttendanceRecordOut, RecordSessionAttendanceIn)
  - wellness.py  (WellnessPreOut, WellnessPostOut, SubmitWellnessPreIn, SubmitWellnessPostIn, ...)
  - execution.py (ExecutionRecordOut, SessionObjectiveOut, LoadChartEntryOut, LoadChartOut)
  - planning.py  (MesocycleOut, MicrocycleOut, ...)
  - communication.py (FeedbackThreadOut, AttentionQueueItemOut, RecommendationOut, ...)
  - eligibility.py (AthleteIneligibilityDeclarationOut, SubmitIneligibilityDeclarationIn)
  __init__.py shim preserva surface pública completa (52 classes re-exportadas)

Fase 5.7 — test_layer_separation.py já verde (nenhuma alteração necessária)

Fase 5.8 — api/errors.py: isinstance fallback para 6 bases semânticas
  (NotFoundError→404, AuthorizationError→403, ConflictError→409,
   PreconditionError→400, StateError→422, DomainValidationError→422)

Tests: 375 passed, 19 skipped
hb verify: exitcode 0
