+++ src/training/schemas/sessions.py
@@ -1,0 +1,120 @@
+from __future__ import annotations
+
+import uuid
+from datetime import datetime
+from decimal import Decimal
+from typing import List, Optional
+
+from ninja import Schema
+
+
+# ---------------------------------------------------------------------------
+# TrainingSession schemas
+# ---------------------------------------------------------------------------
+
+class TrainingSessionOut(Schema):
+    id: uuid.UUID
+    organization_id: uuid.UUID
+    session_at: datetime
+    session_type: str
+    status: str
+    created_by_user_id: uuid.UUID
+    created_at: datetime
+    updated_at: datetime
+    team_id: Optional[uuid.UUID] = None
+    season_id: Optional[uuid.UUID] = None
+    microcycle_id: Optional[uuid.UUID] = None
+    duration_planned_minutes: Optional[int] = None
+    location: Optional[str] = None
+    main_objective: Optional[str] = None
+    secondary_objective: Optional[str] = None
+    planned_load: Optional[int] = None
+    intensity_target: Optional[int] = None
+    session_block: Optional[str] = None
+    notes: Optional[str] = None
+    group_climate: Optional[int] = None
+    standalone: Optional[bool] = None
+    individualization_mode: Optional[str] = None
+    focus_attack_positional_pct: Optional[Decimal] = None
+    focus_defense_positional_pct: Optional[Decimal] = None
+    focus_transition_offense_pct: Optional[Decimal] = None
+    focus_transition_defense_pct: Optional[Decimal] = None
+    focus_attack_technical_pct: Optional[Decimal] = None
+    focus_defense_technical_pct: Optional[Decimal] = None
+    focus_physical_pct: Optional[Decimal] = None
+    phase_focus_defense: Optional[bool] = None
+    phase_focus_attack: Optional[bool] = None
+    phase_focus_transition_offense: Optional[bool] = None
+    phase_focus_transition_defense: Optional[bool] = None
+
+
+class TrainingSessionListOut(Schema):
+    items: List[TrainingSessionOut]
+    next_page_token: Optional[str] = None
+
+
+class CreateTrainingSessionIn(Schema):
+    organization_id: uuid.UUID
+    session_at: datetime
+    session_type: str
+    team_id: Optional[uuid.UUID] = None
+    season_id: Optional[uuid.UUID] = None
+    microcycle_id: Optional[uuid.UUID] = None
+    duration_planned_minutes: Optional[int] = None
+    location: Optional[str] = None
+    main_objective: Optional[str] = None
+    secondary_objective: Optional[str] = None
+    planned_load: Optional[int] = None
+    intensity_target: Optional[int] = None
+    session_block: Optional[str] = None
+    standalone: Optional[bool] = None
+    focus_attack_positional_pct: Optional[Decimal] = None
+    focus_defense_positional_pct: Optional[Decimal] = None
+    focus_transition_offense_pct: Optional[Decimal] = None
+    focus_transition_defense_pct: Optional[Decimal] = None
+    focus_attack_technical_pct: Optional[Decimal] = None
+    focus_defense_technical_pct: Optional[Decimal] = None
+    focus_physical_pct: Optional[Decimal] = None
+    phase_focus_defense: Optional[bool] = None
+    phase_focus_attack: Optional[bool] = None
+    phase_focus_transition_offense: Optional[bool] = None
+    phase_focus_transition_defense: Optional[bool] = None
+
+
+class UpdateTrainingSessionIn(Schema):
+    session_at: Optional[datetime] = None
+    session_type: Optional[str] = None
+    duration_planned_minutes: Optional[int] = None
+    location: Optional[str] = None
+    main_objective: Optional[str] = None
+    secondary_objective: Optional[str] = None
+    planned_load: Optional[int] = None
+    intensity_target: Optional[int] = None
+    session_block: Optional[str] = None
+    standalone: Optional[bool] = None
+    notes: Optional[str] = None
+    focus_attack_positional_pct: Optional[Decimal] = None
+    focus_defense_positional_pct: Optional[Decimal] = None
+    focus_transition_offense_pct: Optional[Decimal] = None
+    focus_transition_defense_pct: Optional[Decimal] = None
+    focus_attack_technical_pct: Optional[Decimal] = None
+    focus_defense_technical_pct: Optional[Decimal] = None
+    focus_physical_pct: Optional[Decimal] = None
+    phase_focus_defense: Optional[bool] = None
+    phase_focus_attack: Optional[bool] = None
+    phase_focus_transition_offense: Optional[bool] = None
+    phase_focus_transition_defense: Optional[bool] = None
+
+
+class TransitionOut(Schema):
+    id: uuid.UUID
+    status: str
+    updated_at: datetime
+
+
+# ---------------------------------------------------------------------------
+# General error schema
+# ---------------------------------------------------------------------------
+
+class ErrorOut(Schema):
+    detail: str
